requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.1
django-allauth==0.49.0
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
whitenoise==6.0.0
